+++ puzzle/class_permuter.cc
@@ -24,10 +24,10 @@ void ClassPermuter::AdvancerBase::AdvanceWithSkip() {
 }
 
 // static
-double ClassPermuter::PermutationCount(const Descriptor* d) {
+int ClassPermuter::PermutationCount(const Descriptor* d) {
   if (d == nullptr) return 0;
 
-  double ret = 1;
+  int ret = 1;
   int value_count = d->Values().size();
   for (int i = 2; i <= value_count; i++) {
     ret *= i;
