+++ puzzle/filtered_solution_permuter.h
@@ -21,7 +21,7 @@ class FilteredSolutionPermuter final : public SolutionPermuter {
     Advancer(const Advancer&) = delete;
     Advancer& operator=(const Advancer&) = delete;
 
-    double position() const override;
+    int64_t position() const override;
     double completion() const override;
 
    private:
@@ -60,7 +60,7 @@ class FilteredSolutionPermuter final : public SolutionPermuter {
     return iterator(absl::make_unique<Advancer>(nullptr));
   }
 
-  double permutation_count() const;
+  int64_t permutation_count() const;
   const ClassPermuter* class_permuter(int class_int) const {
     return class_permuters_[class_int].get();
   }
